added testing.
Still working on bno085 tests. lsm6dso32 and mpu9250 are done for the
most part (found best libraries). Both of them (+ calibratedQuart) have
readme's, please read and add suggestions.

calibratedQuart could be interesting to investigate and try to
understand the maths, a madgwick filter could be a good for our
application.

h3lis is discounted due to it's extremely noisy input (± 0.8g) on each
I2C packet. I think a lsm6dso32 could be a good replacement (32g instead
of 100g), but for the next meeting